implement on project

+++ index.php
@@ -3,6 +3,11 @@ require_once('classes/database.php');
 $con = new database();
 session_start();
 
+if (!isset($_SESSION['username']) || $_SESSION['account_type'] != 0) {
+  header('location:login.php');
+  exit();
+}
+
 if (isset($_POST['delete'])) {
     $id = $_POST['id'];
     if ($con->delete($id)) {
@@ -115,6 +120,11 @@ document.addEventListener('DOMContentLoaded', function() {
       case 'success':
         title = 'Success!';
         text = 'Record is successfully deleted.';
+        icon = 'success';
+        break;
+      case 'login':
+        title = 'Success!';
+        text = 'You are already logged in.';
         icon = 'success';
         break;
       case 'error':
